requirements.txt
requests
beautifulsoup4
slack_sdk
lxml
python-dotenv